# perf(gnolang): improve `print` and `println` performance