--- a/gnovm/pkg/gnolang/native_perf_and_test.go
+++ b/gnovm/pkg/gnolang/native_perf_and_test.go
@@ -53,6 +53,7 @@ func BenchmarkGnoPrintln(b *testing.B) {
 	if pSink == nil {
 		b.Fatal("Benchmark did not run!")
 	}
+	pSink = nil
 }
 
 func TestGnoPrintAndPrintln(t *testing.T) {