--- a/tm2/pkg/bft/rpc/lib/server/write_endpoints_test.go
+++ b/tm2/pkg/bft/rpc/lib/server/write_endpoints_test.go
@@ -6,7 +6,6 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
 
 	types "github.com/gnolang/gno/tm2/pkg/bft/rpc/lib/types"
@@ -27,7 +26,5 @@ func TestWriteListOfEndpoints(t *testing.T) {
 	assert.NoError(t, err)
 	gotResp := string(blob)
 	wantResp := `<html><body><br>Available endpoints:<br><a href="//localhost/d">//localhost/d</a></br><br>Endpoints that require arguments:<br><a href="//localhost/c?s=_&i=_">//localhost/c?s=_&i=_</a></br></body></html>`
-	if diff := cmp.Diff(gotResp, wantResp); diff != "" {
-		t.Fatalf("Mismatch response: got - want +\n%s", diff)
-	}
+	assert.Equal(t, wantResp, gotResp)
 }