This change is the result of seeing print/println being reported in bugs as needing gas metering but on examination, noticed that the code causes a RAM and CPU bloat, so this makes an improvement in all dimensions by invoking:
* io.WriteString(m.Output, segments) instead of m.Output.Write([]byte(segments))
* in cases where `!debug`, we don't need to construct the concatenated 
  string delimited by " " and can instead write the Sprint-ed string 
segments as we encounter them

```shell
$ benchstat before_println.txt after_println3.txt
name          old time/op    new time/op    delta
GnoPrintln-8    1.94ms ± 2%    1.86ms ± 4%   -4.02%  (p=0.001 n=9+9)

name          old alloc/op   new alloc/op   delta
GnoPrintln-8    1.92MB ± 0%    1.81MB ± 0%   -5.92%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
GnoPrintln-8     19.0k ± 0%     16.0k ± 0%  -15.76%  (p=0.002 n=8+10)
```

Fixes #3951